// convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  auth: typeof auth;
  crons: typeof crons;
  files: typeof files;
  http: typeof http;
  pins: typeof pins;
  polar: typeof polar;
  router: typeof router;
}>;
declare const fullApiWithMounts: typeof fullApi;

export declare const api: FilterApi<
  typeof fullApiWithMounts,
  FunctionReference<any, "public">
>;